coverage above 70%
