+++ src/main/java/com/qa/gateway/controller/Controller.java
@@ -13,6 +13,8 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.client.RestTemplate;
 
+import com.qa.gateway.entities.User;
+
 @RestController
 public class Controller {
