+++ src/main/java/com/qa/gateway/entities/User.java
@@ -1,0 +1,78 @@
+package com.qa.gateway.entities;
+
+import java.util.Set;
+
+public class User {
+
+	private long userID;
+	private Set<Booking> bookings;
+
+	private String forename;
+	private String lastname;
+	private String password;
+	private String role;
+	private String email;
+
+	public long getUserID() {
+		return userID;
+	}
+
+	public void setUserID(long userID) {
+		this.userID = userID;
+	}
+
+	public String getForename() {
+		return forename;
+	}
+
+	public void setForename(String forename) {
+		this.forename = forename;
+	}
+
+	public String getLastname() {
+		return lastname;
+	}
+
+	public void setLastname(String lastname) {
+		this.lastname = lastname;
+	}
+
+	public String getPassword() {
+		return password;
+	}
+
+	public void setPassword(String password) {
+		this.password = password;
+	}
+
+	public String getRole() {
+		return role;
+	}
+
+	public void setRole(String role) {
+		this.role = role;
+	}
+
+	public String getEmail() {
+		return email;
+	}
+
+	public void setEmail(String email) {
+		this.email = email;
+	}
+
+	@Override
+	public String toString() {
+		return "User [userID=" + userID + ", forename=" + forename + ", lastname=" + lastname + ", role=" + role
+				+ ", email=" + email + "]";
+	}
+
+	public Set<Booking> getBookings() {
+		return bookings;
+	}
+
+	public void setBookings(Set<Booking> bookings) {
+		this.bookings = bookings;
+	}
+
+}
