+++ src/test/java/com/qa/gateway/entities/BookingEntityTest.java
@@ -1,0 +1,53 @@
+package com.qa.gateway.entities;
+
+import static org.junit.Assert.assertEquals;
+
+import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.context.junit4.SpringRunner;
+
+@RunWith(SpringRunner.class)
+@SpringBootTest
+public class BookingEntityTest {
+	
+	Booking testBook = new Booking();
+
+	@Test
+	public void StartTime() {
+	testBook.setStartTime("a");
+		assertEquals("a", testBook.getStartTime());
+	}
+	
+	@Test
+	public void EndTimeTest() {
+		testBook.setEndTime("a");
+		assertEquals("a", testBook.getEndTime());
+	}
+	
+	@Test
+	public void PeopleTest() {
+		testBook.setNumberOfPeople(1);
+		assertEquals(1, testBook.getNumberOfPeople());
+	}
+	
+	@Test
+	public void DescriptionTest() {
+		testBook.setDescription("a");
+		assertEquals("a", testBook.getDescription());
+	}
+	
+	@Test
+	public void TitleTest() {
+		testBook.setTitle("a");
+		assertEquals("a", testBook.getTitle());
+	}
+	
+	@Test
+	public void StringTest() {
+		assertEquals("Booking [bookingID=0, startTime=null, endTime=null, numberOfPeople=1, title=null, description=null]", testBook.toString());
+	}
+	
+	
+
+}
